backend/agents/requirements.txt
google-adk
httpx
python-dotenv
litellm